Renamed `functions.py` to `common.py`
+ Moved action functions to separate file
