+++ src/backend.py
@@ -1,5 +1,5 @@
 from fastapi.responses import FileResponse, JSONResponse, RedirectResponse
-from functions import logger, SOURCE_DIR, log_file_path
+from common import logger, SOURCE_DIR, log_file_path
 from fastapi import FastAPI, Request, Response, status
 from spotipy.oauth2 import SpotifyClientCredentials
 from dotenv import load_dotenv, set_key, unset_key
